requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-cors-headers==4.3.1
django-jazzmin==2.6.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.6
nose==1.3.7
pillow==10.2.0
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.10.0
urllib3==2.2.1
